✨ feat: Introduce realme_macros builder! macro
Added a builder! macro in realme_macros for streamlined Realme construction.  Upgraded syn and quote dependencies and introduced an adaptor module.  Refactored file adaptor logic.
