+++ Cargo.toml
@@ -11,7 +11,7 @@ homepage = "https://github.com/vainjoker/realme"
 description = "A flexible and extensible configuration management library for Rust, designed to simplify the process of loading and managing configuration settings from various sources."
 
 [features]
-default = ["env", "macros"]
+default = ["env", "macros", "toml"]
 full = [
   "cmd",
   "env",
