+++ realme_macros/Cargo.toml
@@ -15,7 +15,6 @@ proc-macro = true
 bench = false
 
 [dependencies]
-syn = { version = "2.0.8", features = ["full"] }
-quote = "1.0.9"
-proc-macro2 = "1.0.69"
-darling = "0.20.10"
+syn = { version = "2.0.90", features = ["extra-traits"] }
+quote = "1.0"
+proc-macro2 = "1.0"
